+++ shortbread/src/main/java/shortbread/internal/MethodShortcuts.java
@@ -1,0 +1,10 @@
+package shortbread.internal;
+
+import android.app.Activity;
+
+public interface MethodShortcuts<T extends Activity> {
+
+    Class getActivityClass();
+
+    void callMethodShortcut(Activity activity, String methodName);
+}