+++ shortbread/src/test/java/com/example/TestMethodShortcutActivity.java
@@ -1,0 +1,7 @@
+package com.example;
+
+import android.app.Activity;
+
+public class TestMethodShortcutActivity extends Activity {
+
+}