+++ shortbread/src/test/resources/validation/NoParameterlessMethod.java
@@ -7,7 +7,7 @@ import shortbread.Shortcut;
 public class NoParameterlessMethod extends Activity {
 
     @Shortcut(id = "ID")
-    public void startActivity(Object parameter) {
+    void startActivity(Object parameter) {
 
     }
 }